Adicionando informação de preço antigo em promoção e personalizando card sem a imagem

+++ app/Service.php
@@ -7,7 +7,7 @@ use Illuminate\Database\Eloquent\Model;
 class Service extends Model
 {
     function getPriceAttribute(){
-        return number_format($this->attributes['price'], 2, ',', '.');
+        return (float) number_format($this->attributes['price'], 2, ',', '.');
     }
 
 }
